* 'feature_dynamic_dims_cuda' of https://github.com/Tencent/TNN: (301 commits)
  [CUDA][UPD][FIX]update tensorrt network;fix cuda einsum
  [ARM] fix matmul get weights error for two inputs
  fix network md5 missing && default network call layer reshape
  bypass conv7x7 following concat/conv with dialation=4 case
  [ARM] skip fp16 fc unittest for large input size
  [ARM] support reload to packed const blobs
  [ARM][BUG] fix binary crash when forward multiple times
  [UNIT TEST][REFACTOR] add check data type function
  [DOC] update development guide for x86
  [DEVICE][CUDA] fix compile error
  [OPECNL] fix innerproduct && concat unit test in opencl && skip reduce test for not keep dims
  [OPENCL] skip unit test of op not support 5/6 dims
  [OPENCL] add mat mul layer in opencl
  [OPENVINO] 1.update build_x86_linux.sh. 2.fix bertsquad10 reshape error
  [ARM] refactor arm binary op to support fp16 and dynamic shape
  [UNIT TEST][PATCH] remove some conv channel check
  [ARM] optimize pack for half blob
  [CONVWETER2TNN][BUG]1. specify input info support generate random int32 ;
  [METAL][FIX] fix compile error
  [UNIT TEST][UNARY] change for unary test
  ...

+++ source/tnn/memory_manager/blob_2d_memory_pool.h
@@ -31,7 +31,7 @@ private:
     virtual BlobMemory* CreateBlobMemory(int use_count, BlobMemorySizeInfo& size_info) override;
     virtual BlobMemoryNode* GetBlobMemoryNodeListHeader(DataType data_type) override;
     virtual void SetBlobMemoryNodeListHeader(DataType data_type, BlobMemoryNode* new_header) override;
-    virtual int ResolveBlobMemoryNodeBytesDiff(BlobMemorySizeInfo& size_info, BlobMemoryNode* node) override;
+    virtual int64_t ResolveBlobMemoryNodeBytesDiff(BlobMemorySizeInfo& size_info, BlobMemoryNode* node) override;
     // extract the closest BlobMemoryNode from BlobMemoryNode list for 2D memory
     virtual BlobMemoryNode* ExtractNearestBlobMemoryNode(BlobMemorySizeInfo& size_info) override;
     std::map<DataType, BlobMemoryNode*> blob_memory_list_header_map_;
